requirements.txt
arrow==1.2.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
bfg==0.1.0
Django==4.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.3.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
iniconfig==2.0.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
tzdata==2023.3
